+++ src/util/util.c
@@ -1,11 +1,11 @@
 #include <stdint.h>
+#include <stddef.h>
 
 #include "util/util.h"
 
-
-unsigned long util_fnv1a_hash(const char* key)
+unsigned long long util_fnv1a_hash(const char* key)
 {
-    unsigned long hash = FNV_OFFSET_BASIS;
+    unsigned long long hash = FNV_OFFSET_BASIS;
     while (*key)
     {
         hash *= FNV_PRIME;
